Also returns previous values for municipalities stats
